+++ src/components/useThree/mainCanvas.tsx
@@ -2,8 +2,15 @@ import { OrbitControls } from "@react-three/drei";
 import { Canvas } from "@react-three/fiber";
 import { Moon } from "./moon";
 import { Universe } from "./universe";
+import type { MoonquakeData } from "@/type";
 
-export const MainCanvas = () => {
+type Props = {
+  moonquakeData: MoonquakeData[];
+  setIsMap: React.Dispatch<React.SetStateAction<boolean>>;
+};
+export const MainCanvas = (props: Props) => {
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  const { moonquakeData, setIsMap } = props;
   return (
     <Canvas
       camera={{
@@ -17,7 +24,7 @@ export const MainCanvas = () => {
       <directionalLight position={[1, 1, 1]} intensity={0.8} />
       <ambientLight args={[0xffffff]} intensity={0.5} />
       <OrbitControls minDistance={250} maxDistance={500} />
-      <Moon />
+      <Moon moonquakeData={moonquakeData} />
       <Universe />
     </Canvas>
   );
